+++ pubspec.yaml
@@ -30,6 +30,7 @@ environment:
 # the latest version available on pub.dev. To see which dependencies have newer
 # versions available, run `flutter pub outdated`.
 dependencies:
+  async: ^2.10.0
   auto_route: ^6.3.0
   bip32: ^2.0.0
   bip39: ^1.0.6
@@ -41,15 +42,15 @@ dependencies:
   flutter:
     sdk: flutter
   flutter_js: ^0.6.0
+  flutter_mobx: ^2.0.6+5
   flutter_screenutil: ^5.7.0
   flutter_svg: ^2.0.5
-  freezed_annotation: ^2.2.0  
+  freezed_annotation: ^2.2.0
   get_it: ^7.3.0
   hex: ^0.2.0
   http2: 2.0.1
   js: ^0.6.5
   mobx: ^2.1.4
-  flutter_mobx: ^2.0.6+5  
   overlay_support: ^2.1.0
   pinenacl: ^0.5.1
   slip_0010_ed25519: ^3.1.0
